--- a/desert-cats/src/main/scala/io/github/vigoo/desert/catssupport/codecs.scala
+++ b/desert-cats/src/main/scala/io/github/vigoo/desert/catssupport/codecs.scala
@@ -2,27 +2,23 @@ package io.github.vigoo.desert
 
 import cats.Order
 import cats.data.{NonEmptyList, NonEmptyMap, NonEmptySet, Validated}
-import cats.instances.either._
-import cats.syntax.flatMap._
-
-import io.github.vigoo.desert._
 import io.github.vigoo.desert.custom._
+import io.github.vigoo.desert.internal.{DeserializationContext, SerializationContext}
 
 package object catssupport {
   // Cats specific codecs
 
   implicit def validatedCodec[E: BinaryCodec, A: BinaryCodec]: BinaryCodec[Validated[E, A]] =
     new BinaryCodec[Validated[E, A]] {
-      override def deserialize(): Deser[Validated[E, A]] =
-        for {
-          isValid <- read[Boolean]()
-          result  <- if (isValid) read[A]().map(Validated.Valid.apply) else read[E]().map(Validated.Invalid.apply)
-        } yield result
+      override def deserialize()(implicit ctx: DeserializationContext): Validated[E, A] = {
+        val isValid = read[Boolean]()
+        if (isValid) Validated.Valid(read[A]()) else Validated.Invalid(read[E]())
+      }
 
-      override def serialize(value: Validated[E, A]): Ser[Unit] =
+      override def serialize(value: Validated[E, A])(implicit context: SerializationContext): Unit =
         value match {
-          case Validated.Invalid(value) => write(false) *> write(value)
-          case Validated.Valid(value)   => write(true) *> write(value)
+          case Validated.Valid(a)   => write(true); write(a)
+          case Validated.Invalid(e) => write(false); write(e)
         }
     }
 
@@ -30,14 +26,16 @@ package object catssupport {
     val inner = listCodec[A]
     BinaryCodec.from(
       inner.contramap(_.toList),
-      () =>
-        inner.deserialize().flatMap { list =>
+      new BinaryDeserializer[NonEmptyList[A]] {
+        override def deserialize()(implicit ctx: DeserializationContext): NonEmptyList[A] = {
+          val list = inner.deserialize()
           NonEmptyList.fromList(list) match {
-            case Some(value) => finishDeserializerWith(value)
+            case Some(value) => value
             case None        =>
               failDeserializerWith(DesertFailure.DeserializationFailure("Non empty list is serialized as empty", None))
           }
         }
+      }
     )
   }
 
@@ -45,14 +43,16 @@ package object catssupport {
     val inner = sortedSetCodec[A]
     BinaryCodec.from(
       inner.contramap(_.toSortedSet),
-      () =>
-        inner.deserialize().flatMap { set =>
+      new BinaryDeserializer[NonEmptySet[A]] {
+        override def deserialize()(implicit ctx: DeserializationContext): NonEmptySet[A] = {
+          val set = inner.deserialize()
           NonEmptySet.fromSet(set) match {
-            case Some(value) => finishDeserializerWith(value)
+            case Some(value) => value
             case None        =>
               failDeserializerWith(DesertFailure.DeserializationFailure("Non empty set is serialized as empty", None))
           }
         }
+      }
     )
   }
 
@@ -60,14 +60,16 @@ package object catssupport {
     val innner = sortedMapCodec[K, V]
     BinaryCodec.from(
       innner.contramap(_.toSortedMap),
-      () =>
-        innner.deserialize().flatMap { map =>
+      new BinaryDeserializer[NonEmptyMap[K, V]] {
+        override def deserialize()(implicit ctx: DeserializationContext): NonEmptyMap[K, V] = {
+          val map = innner.deserialize()
           NonEmptyMap.fromMap(map) match {
-            case Some(value) => finishDeserializerWith(value)
+            case Some(value) => value
             case None        =>
               failDeserializerWith(DesertFailure.DeserializationFailure("Non empty map is serialized as empty", None))
           }
         }
+      }
     )
   }
 }