--- a/desert-core/src/main/scala/io/github/vigoo/desert/RegisteredType.scala
+++ b/desert-core/src/main/scala/io/github/vigoo/desert/RegisteredType.scala
@@ -2,11 +2,12 @@ package io.github.vigoo.desert
 
 import io.github.vigoo.desert.TypeRegistry.RegisteredTypeId
 import io.github.vigoo.desert.custom.failSerializerWith
+import io.github.vigoo.desert.internal.{SerializationContext, SerializationEnv, SerializerState}
 
 import scala.util.{Failure, Success, Try}
 
 final case class RegisteredType[T](id: RegisteredTypeId, codec: BinaryCodec[T], cls: Class[_]) {
-  def serialize(value: Any): Ser[Unit] =
+  def serialize(value: Any)(implicit ctx: SerializationContext): Unit =
     Try(value.asInstanceOf[T]) match {
       case Success(upcasted)  => codec.serialize(upcasted)
       case Failure(exception) =>