--- a/desert-core/src/main/scala/io/github/vigoo/desert/custom/BinarySerializerOps.scala
+++ b/desert-core/src/main/scala/io/github/vigoo/desert/custom/BinarySerializerOps.scala
@@ -1,6 +1,5 @@
 package io.github.vigoo.desert.custom
 
-import io.github.vigoo.desert.internal.SerializationContext
 import io.github.vigoo.desert.internal.SerializerState.{RefAlreadyStored, RefIsNew, StoreRefResult, StoreStringResult}
 import io.github.vigoo.desert.{BinaryCodec, BinarySerializer, DesertException, DesertFailure}