--- a/desert-core/src/main/scala/io/github/vigoo/desert/custom/DeserializationContext.scala
+++ b/desert-core/src/main/scala/io/github/vigoo/desert/custom/DeserializationContext.scala
@@ -1,0 +1,5 @@
+package io.github.vigoo.desert.custom
+
+import io.github.vigoo.desert.internal.{DeserializationEnv, SerializerState}
+
+final case class DeserializationContext(env: DeserializationEnv, state: SerializerState)