--- a/desert-core/src/main/scala/io/github/vigoo/desert/custom/SerializationContext.scala
+++ b/desert-core/src/main/scala/io/github/vigoo/desert/custom/SerializationContext.scala
@@ -1,0 +1,5 @@
+package io.github.vigoo.desert.custom
+
+import io.github.vigoo.desert.internal.{SerializationEnv, SerializerState}
+
+final case class SerializationContext(env: SerializationEnv, state: SerializerState)