… benchmarks/src/main/scala/io/github/vigoo/desert/benchmarks/GenericComplexSerializationZioSchemaBenchmark.scala
… benchmarks/src/main/scala/io/github/vigoo/desert/benchmarks/LargeCompressedArraySerializationBenchmark.scala
… desert-core/src/main/scala/io/github/vigoo/desert/custom/pure/PureBinarySerializerOps.scala
… desert-core/src/main/scala/io/github/vigoo/desert/custom/pure/package.scala
… desert-core/src/main/scala/io/github/vigoo/desert/internal/AdtCodec.scala
… desert-core/src/main/scala/io/github/vigoo/desert/internal/DeserializationContext.scala
… desert-core/src/main/scala/io/github/vigoo/desert/internal/JavaStreamBinaryInput.scala
… desert-core/src/main/scala/io/github/vigoo/desert/internal/JavaStreamBinaryOutput.scala
… desert-core/src/main/scala/io/github/vigoo/desert/internal/OptionBinaryCodec.scala
… desert-core/src/main/scala/io/github/vigoo/desert/internal/PureSerializerState.scala
… desert-core/src/main/scala/io/github/vigoo/desert/internal/SerializationContext.scala
… desert-core/src/main/scala/io/github/vigoo/desert/internal/SerializerState.scala
… desert-core/src/main/scala/io/github/vigoo/desert/package.scala
… desert-core/src/test/scala/io/github/vigoo/desert/JavaStreamInputOutputSpec.scala
… desert-core/src/test/scala/io/github/vigoo/desert/ReferenceTrackingSpec.scala
… desert-shapeless/src/main/scala/io/github/vigoo/desert/shapeless/GenericBinaryCodec.scala
… desert-shapeless/src/main/scala/io/github/vigoo/desert/shapeless/UnwrappedBinaryCodec.scala
… desert-shapeless/src/test/scala/io/github/vigoo/desert/shapeless/DefaultTypeRegistrySpec.scala
… desert-shapeless/src/test/scala/io/github/vigoo/desert/shapeless/StringDeduplicationSpec.scala
… desert-zio-schema/src/main/scala/io/github/vigoo/desert/zioschema/DerivedBinaryCodec.scala
… desert-zio-schema/src/test/scala/io/github/vigoo/desert/zioschema/DefaultTypeRegistrySpec.scala
… desert-zio-schema/src/test/scala/io/github/vigoo/desert/zioschema/StringDeduplicationSpec.scala
… desert-zio/src/main/scala/io/github/vigoo/desert/ziosupport/Codecs.scala
… desert-zio/src/main/scala/io/github/vigoo/desert/ziosupport/custom/package.scala
… docs/docs/docs/codecs.md
… docs/docs/docs/evolution.md
… docs/docs/docs/index.md
… docs/docs/docs/input-output.md
… docs/docs/docs/release-notes.md
… project/TupleCodecGenerator.scala