old: 

```
[info] GenericComplexSerializationBenchmark.deserializeComplexDataStructure            N/A   34253      12  thrpt   25       0.081 ±   0.001  ops/ms
[info] GenericComplexSerializationBenchmark.serializeComplexDataStructure              N/A   34253      12  thrpt   25       0.094 ±   0.001  ops/ms
[info] JavaStreamBinaryInputBenchmark.readByte                                         N/A     N/A     N/A  thrpt   50   34317.703 ±  75.847  ops/ms
[info] JavaStreamBinaryInputBenchmark.readBytes                                        N/A     N/A     N/A  thrpt   50    1544.846 ±   4.471  ops/ms
[info] JavaStreamBinaryInputBenchmark.readDouble                                       N/A     N/A     N/A  thrpt   50   30458.934 ±  57.228  ops/ms
[info] JavaStreamBinaryInputBenchmark.readFloat                                        N/A     N/A     N/A  thrpt   50   18068.088 ±  80.951  ops/ms
[info] JavaStreamBinaryInputBenchmark.readInt                                          N/A     N/A     N/A  thrpt   50   17877.505 ±  70.633  ops/ms
[info] JavaStreamBinaryInputBenchmark.readLong                                         N/A     N/A     N/A  thrpt   50   30302.676 ±  60.778  ops/ms
[info] JavaStreamBinaryInputBenchmark.readShort                                        N/A     N/A     N/A  thrpt   50   27781.025 ±  47.899  ops/ms
[info] JavaStreamBinaryInputBenchmark.readVarInt                                       N/A     N/A     N/A  thrpt   50   31095.668 ± 162.525  ops/ms
[info] JavaStreamBinaryOutputBenchmark.writeByte                                       N/A     N/A     N/A  thrpt   50  101327.423 ± 438.228  ops/ms
[info] JavaStreamBinaryOutputBenchmark.writeBytes                                      N/A     N/A     N/A  thrpt   50     611.048 ±  16.563  ops/ms
[info] JavaStreamBinaryOutputBenchmark.writeDouble                                     N/A     N/A     N/A  thrpt   50   60591.642 ± 169.784  ops/ms
[info] JavaStreamBinaryOutputBenchmark.writeInt                                        N/A     N/A     N/A  thrpt   50   42407.368 ± 227.795  ops/ms
[info] JavaStreamBinaryOutputBenchmark.writeLong                                       N/A     N/A     N/A  thrpt   50   60202.621 ± 160.301  ops/ms
[info] JavaStreamBinaryOutputBenchmark.writeShort                                      N/A     N/A     N/A  thrpt   50   76707.602 ± 353.615  ops/ms
[info] JavaStreamBinaryOutputBenchmark.writeVarInt                                     N/A     N/A     N/A  thrpt   50   53722.409 ± 296.000  ops/ms
[info] LargeCompressedArraySerializationBenchmark.deserializeLargeCompressedArray        1     N/A     N/A  thrpt   50       1.025 ±   0.002  ops/ms
[info] LargeCompressedArraySerializationBenchmark.serializeLargeCompressedArray          1     N/A     N/A  thrpt   50       0.360 ±   0.001  ops/ms
```

new:

```
[info] GenericComplexSerializationBenchmark.deserializeComplexDataStructure            N/A   34253      12  thrpt   25       1.220 ±    0.013  ops/ms
[info] GenericComplexSerializationBenchmark.serializeComplexDataStructure              N/A   34253      12  thrpt   25       1.378 ±    0.008  ops/ms
[info] GenericComplexSerializationZioSchemaBenchmark.deserializeComplexDataStructure   N/A   34253      12  thrpt   25       0.859 ±   0.003  ops/ms
[info] GenericComplexSerializationZioSchemaBenchmark.serializeComplexDataStructure     N/A   34253      12  thrpt   25       1.107 ±   0.014  ops/ms
[info] JavaStreamBinaryInputBenchmark.readByte                                         N/A     N/A     N/A  thrpt   50   38876.851 ±  109.319  ops/ms
[info] JavaStreamBinaryInputBenchmark.readBytes                                        N/A     N/A     N/A  thrpt   50    1569.202 ±   11.880  ops/ms
[info] JavaStreamBinaryInputBenchmark.readDouble                                       N/A     N/A     N/A  thrpt   50   35824.478 ±  105.515  ops/ms
[info] JavaStreamBinaryInputBenchmark.readFloat                                        N/A     N/A     N/A  thrpt   50   19203.590 ±   21.651  ops/ms
[info] JavaStreamBinaryInputBenchmark.readInt                                          N/A     N/A     N/A  thrpt   50   19273.876 ±   23.784  ops/ms
[info] JavaStreamBinaryInputBenchmark.readLong                                         N/A     N/A     N/A  thrpt   50   35968.912 ±  128.084  ops/ms
[info] JavaStreamBinaryInputBenchmark.readShort                                        N/A     N/A     N/A  thrpt   50   29947.543 ±   54.189  ops/ms
[info] JavaStreamBinaryInputBenchmark.readVarInt                                       N/A     N/A     N/A  thrpt   50   41992.874 ±   68.635  ops/ms
[info] JavaStreamBinaryOutputBenchmark.writeByte                                       N/A     N/A     N/A  thrpt   50  100747.696 ± 1240.757  ops/ms
[info] JavaStreamBinaryOutputBenchmark.writeBytes                                      N/A     N/A     N/A  thrpt   50     633.550 ±    6.088  ops/ms
[info] JavaStreamBinaryOutputBenchmark.writeDouble                                     N/A     N/A     N/A  thrpt   50   63718.069 ±  835.505  ops/ms
[info] JavaStreamBinaryOutputBenchmark.writeInt                                        N/A     N/A     N/A  thrpt   50   42344.990 ±  763.297  ops/ms
[info] JavaStreamBinaryOutputBenchmark.writeLong                                       N/A     N/A     N/A  thrpt   50   68774.549 ±  151.205  ops/ms
[info] JavaStreamBinaryOutputBenchmark.writeShort                                      N/A     N/A     N/A  thrpt   50   87374.803 ±  144.227  ops/ms
[info] JavaStreamBinaryOutputBenchmark.writeVarInt                                     N/A     N/A     N/A  thrpt   50   52840.371 ±  295.112  ops/ms
[info] LargeCompressedArraySerializationBenchmark.deserializeLargeCompressedArray        1     N/A     N/A  thrpt   50       1.024 ±    0.003  ops/ms
[info] LargeCompressedArraySerializationBenchmark.serializeLargeCompressedArray          1     N/A     N/A  thrpt   50       0.365 ±    0.001  ops/ms
```